Rename: reviews, mentors 관련 컴포넌트 파일 정리 (#33)
